requirements.txt
Django==4.2.13
djangorestframework==3.15.2
python-decouple==3.8
requests==2.32.3
celery==5.4.0
